+++ tests/testthat/test-compile_pseudo_pop.R
@@ -2,16 +2,16 @@ test_that("Compiling pseudo pop works as expected.", {
 
   set.seed(509)
   m_d <- generate_syn_data(sample_size = 100)
-  data_with_gps_1 <- estimate_gps(m_d$Y,
-                                  m_d$treat,
-                                  m_d[c("cf1","cf2","cf3","cf4","cf5","cf6")],
-                                  pred_model = "sl",
-                                  internal_use = TRUE,
-                                  sl_lib = c("m_xgboost"))
+  gps_obj <- estimate_gps(m_d$Y,
+                          m_d$treat,
+                          m_d[c("cf1","cf2","cf3","cf4","cf5","cf6")],
+                          pred_model = "sl",
+                          internal_use = TRUE,
+                          sl_lib = c("m_xgboost"))
 
 
   # Wrong ci_appr
-  expect_error(compile_pseudo_pop(dataset=data_with_gps_1, ci_appr="grounding",
+  expect_error(compile_pseudo_pop(dataset=gps_obj$dataset, ci_appr="grounding",
                                    gps_model = "parametric", bin_seq,
                                    nthread = 1,
                                    trim_quantiles = c(0.01, 0.99),
@@ -20,9 +20,9 @@ test_that("Compiling pseudo pop works as expected.", {
 
   #
   set.seed(509)
-  pseudo_pop_1 <- compile_pseudo_pop(dataset=data_with_gps_1,
-                                     ci_appr="matching",
-                                     gps_model="parametric",
+  pseudo_pop_1 <- compile_pseudo_pop(data_obj = gps_obj,
+                                     ci_appr = "matching",
+                                     gps_model = "parametric",
                                      bin_seq = NULL,
                                      nthread = 1,
                                      trim_quantiles = c(0.01, 0.99),
@@ -44,7 +44,10 @@ test_that("Compiling pseudo pop works as expected.", {
 
   set.seed(934)
   data <- list(pseudo_pop_weight_test[, !c("ipw")])
-  pseudo_pop_2 <- compile_pseudo_pop(dataset=data,
+  obj <- list()
+  class(obj) <- "cgps_gps"
+  obj$dataset <- data[[1]]
+  pseudo_pop_2 <- compile_pseudo_pop(data_obj = obj,
                                      ci_appr="weighting",
                                      gps_model="parametric",
                                      bin_seq = NULL,